Test al Controller. Página de error. Test pasa (GREEN)

+++ src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -37,6 +37,11 @@ public class GradebookController {
 
 	@GetMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable int id, Model m) {
+
+		if (!studentService.checkIfStudentIsNull(id)) {
+			return "error";
+		}
+
 		studentService.deleteStudent(id);
 
 		Iterable<CollegeStudent> collegeStudents = studentService.getGradebook();
